refactor: move prompts & events to chat utils folder
