+++ packages/server/src/index.ts
@@ -1,5 +1,5 @@
-export * from "./events";
-export * from "./prompts";
 export * from "./server";
 export * from "./types";
+export * from "./utils/events";
 export { generateEventComponent } from "./utils/gen-ui";
+export * from "./utils/prompts";
