+++ packages/server/src/utils/prompts.ts
